Access to original constructor in mocks
Neccessarily mocks override `__construct`, however it's sometimes
desirable to be able to bootstrap a mock using the code within. This
commit introduces the `__constructor` method, which exists for the
purpose of being able to run the original `__construct` code in mocks.

This also facilitates the late calling of `__construct`, i.e. after an
instance has been created.