Fix addition of new methods to mocks:
In 1.1.0, you could add new methods when generating a mock. This was
useful when testing a class you've yet to write, and also for dealing
with magic methods.

Code comments suggest this feature is still supported, but the code
errors when attempting to build the signature from the reflection, since
the method doesn't exist.

This fix adds support for generating new, arg-less methods to full and
partial mocks.